LPS-180293 Rename ExceptionMapper classes, so they contain the REST entity name (Site) instead of the internal entity name (Group)
