+++ modules/apps/headless/headless-site/headless-site-impl/src/main/java/com/liferay/headless/site/internal/jaxrs/exception/mapper/NoSuchSiteExceptionMapper.java
@@ -14,9 +14,13 @@
 
 package com.liferay.headless.site.internal.jaxrs.exception.mapper;
 
-import com.liferay.portal.kernel.exception.GroupKeyException;
+import com.liferay.portal.kernel.exception.NoSuchGroupException;
+import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.vulcan.jaxrs.exception.mapper.BaseExceptionMapper;
 import com.liferay.portal.vulcan.jaxrs.exception.mapper.Problem;
+
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 import javax.ws.rs.core.Response;
 import javax.ws.rs.ext.ExceptionMapper;
@@ -24,7 +28,7 @@ import javax.ws.rs.ext.ExceptionMapper;
 import org.osgi.service.component.annotations.Component;
 
 /**
- * Converts any {@code GroupKeyExceptionMapper} to a {@code 400} error.
+ * Converts any {@code NoSuchGroupExceptionMapper} to a {@code 404} error.
  *
  * @author Rubén Pulido
  */
@@ -32,16 +36,41 @@ import org.osgi.service.component.annotations.Component;
 	property = {
 		"osgi.jaxrs.application.select=(osgi.jaxrs.name=Liferay.Headless.Site)",
 		"osgi.jaxrs.extension=true",
-		"osgi.jaxrs.name=Liferay.Headless.Site.GroupKeyExceptionMapper"
+		"osgi.jaxrs.name=Liferay.Headless.Site.NoSuchSiteExceptionMapper"
 	},
 	service = ExceptionMapper.class
 )
-public class GroupKeyExceptionMapper
-	extends BaseExceptionMapper<GroupKeyException> {
+public class NoSuchSiteExceptionMapper
+	extends BaseExceptionMapper<NoSuchGroupException> {
 
 	@Override
-	protected Problem getProblem(GroupKeyException groupKeyException) {
-		return new Problem(Response.Status.BAD_REQUEST, "Site key is invalid");
+	protected Problem getProblem(NoSuchGroupException noSuchGroupException) {
+		return new Problem(
+			Response.Status.NOT_FOUND, _getTitle(noSuchGroupException));
 	}
+
+	private String _getTitle(NoSuchGroupException noSuchGroupException) {
+		String title = "No site exists";
+
+		if (noSuchGroupException.getMessage() == null) {
+			return title;
+		}
+
+		Matcher matcher = _pattern.matcher(noSuchGroupException.getMessage());
+
+		if (!matcher.matches()) {
+			return title;
+		}
+
+		String siteKey = matcher.group(1);
+
+		if (Validator.isNotNull(siteKey)) {
+			title = title + " for site key " + siteKey;
+		}
+
+		return title;
+	}
+
+	private static final Pattern _pattern = Pattern.compile(".+groupKey=(.+)}");
 
 }
